Verified cleanup of real review minors (behaviour-preserving)
Apply the minor findings from the final review that were independently
verified as genuine AND safe (no routing-behaviour change):

- RoutingParamCollector: invalidate non-positive roundTripLength so it
  can't yield a zero/negative searchRadius (length / 2π); falls back to
  the roundTripDistance path, mirroring the roundTripPoints guard.
- ReuseClassifier: accumulate stem/spur reuse as double and round once
  where surfaced, instead of truncating each stretch into an int. These
  values are disclosure-only (not classification thresholds), so output
  is unchanged bar sub-meter rounding.
- GreedyRoundTripPlanner: update totalDistance after appending the return
  leg so the FINAL-TRACK diagnostic reflects the closed loop.
- IsochroneCandidateProvider: DEDUPE_GRANULARITY long -> int to match
  GraphNativeCandidateProvider; quotient/packing identical (value fits int).

Deliberately NOT applied (verified false-alarm or behaviour-changing):
LoopQualityMetrics edge-key rehash (feeds RouteChoiceScore candidate
selection — not behaviour-preserving), NaN guard (already guarded),
copyRequestFields keyValues copy and field-resets (single-use engine),
VisitedEdgeStore assert (callers already guard), RouteChoiceScore %n
(literal already contains %n — works correctly).

gradlew build green; ReuseClassifier/Isochrone suites pass.

+++ brouter-core/src/main/java/btools/router/IsochroneCandidateProvider.java
@@ -28,7 +28,7 @@ final class IsochroneCandidateProvider implements RoundTripCandidateProvider {
   private static final double STEP_WINDOW_LOW = 0.5;
   private static final double STEP_WINDOW_HIGH = 1.6;
   /** Dedupe granularity in ilon units (~7m at lat 50°N — within snap range). */
-  private static final long DEDUPE_GRANULARITY = 100;
+  private static final int DEDUPE_GRANULARITY = 100;
   /** Below this filtered-pool size, we relax the hits<3 filter (matches buildCandidateProvider gate). */
   private static final int MIN_DIVERSITY_BEFORE_RELAX = 6;
   /** Minimum distinct angular buckets the filtered pool must span. */
